# swagger.yaml
...
swagger: '2.0'
info:
  title: IoT Device Repository
  termsOfService: http://swagger.io/terms/
  description: 'The IOT Device Repository provides metadata for all IOT devices managed on the platform, their device classes and the associated data structures for the user.'
  version: '1.0'
  contact:
    name: sh20xyqi
    email: sh20xyqi@studserv.uni-leipzig.de
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
schemes:
  - https
host: api.sepl.infai.org
basePath: /iot-device-repo
produces:
  - text/plain
...
paths:
...
  /valueType/{id}:
    delete:
        tags:
          - otherStructs
        summary:
          Deletes valueType if no dependent valueTypes or deviceTypes exist.
        description:
          ''
        produces:
          - application/json
        parameters:
          - name: id
            in: path
            description: Identifies the valueType to be deleted.
            required: true
            type: string
        responses:
          200:
            description: valueType deleted.
            schema:
              type: string
          500:
            description: Returns an ErrorMessage containing further information about the specific error.
            schema:
              $ref: '#/definitions/ErrorMessage'
    get:
      tags:
        - otherStructs
      summary:
        Returns valueType by Id.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: Identifies the to be returned value type.
          required: true
          type: string
      responses:
        200:
          description: Returns the valueType behind the passed Id.
          schema:
            $ref: '#/definitions/ValueType'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - ValueType can not be found in database.
          schema:
            $ref: '#/definitions/ErrorMessage'
  /valueTypes/{limit}/{offset}:
    get:
      tags:
        - otherStructs
      summary:
        Returns value types within a limit and how much elements had to be skipped.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: limit
          in: path
          description: Restricts the set of returnable elements to the passed limit.
          required: true
          type: integer
        - name: offset
          in: path
          description: Specifies how many elements of the set are skipped.
          required: true
          type: integer
      responses:
        200:
          description: Returns a sorted list of value types.
          schema:
            type: array
            items:
              $ref: '#/definitions/ValueType'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - A specific limit can not be found; a specific offset can not be found; a list of ValueTypes can not be returned.
          schema:
            $ref: '#/definitions/ErrorMessage'
  #Everything that belongs to search
  #Everything that belongs to search/search.go
...
definitions:
...
  FieldType:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
      type:
        type: object
        format: '#/definitions/ValueType'
    required:
      - id
      - name
      - type
  ValueType:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
      description:
        type: string
        format: string
      base_type:
        type: string
        format: string
      fields:
        type: array
        items:
          $ref: '#/definitions/FieldType'
      literal:
        type: string
        format: string
    required:
      - id
      - name
      - description
      - base_type
      - fields
      - literal
...
  ErrorMessage:
    type: object
    properties:
      status_code:
        type: integer
        format: int32
      message:
        type: string
        format: string
      error_code:
        type: string
        format: string
      detail:
        type: array
        items:
          type: string
    required:
      - status_code
      - message
      - error_code
      - detail
...